Use self written LSP config without `lsp-zero`
I got tired of not being able to override all the settings (probably because I don't know lua very well, but still), which led me to dig into the plugin's source code and change it there. Therefore, I think it's better to manually configure the LSP, CMP, and Mason settings.