Add check for 'nvim' environment variable in 'esc' script
The commit adds a check for the presence of the 'nvim' environment variable in the 'esc' script. If 'nvim' is not set, the script now displays an error message and exits with a non-zero status code. This helps to ensure that the script will work correctly even if 'nvim' is not set or set to an invalid value.